Disable package installation inside config of provider "virtualbox"
The configuration of the provider is executed regardless of whether the provider
was selected or not. This means that any other code in this configuration block
is executed even when another provider is specified.

In this specific case several VirtualBox-specific packages were installed in the
new VM even when provider "libvirt" was used. This is not what you want so this
commit comments the relevant lines.