Touch curl link created by Stow
Stow doesn't update time stamps on links. Now assume the underlying package in
the Stow directory is updated, then the link will be older than its target. The
target that stows a package will execute but not update the time stamp on the
link. As a result, the target will always execute.