Update curl target to allow deletion of downloaded archive
This commit uses another approach to allow deletion of builds and build
dependencies. This approach does not use a no-retrieve file and is to be
preferred to the one that does.

This commit also documents the curl target so it can serve as a reference target
for packages and especially packages that require a configure and build step.